Added better AI for enemy. Added UI and added the ability to "kill" enemies
The enemies now patrol from 2 points, a left and a right patrol point.
There is now a score system that once you kill an enemy it goes up by 100.
Enemies are killed by jumping on top of them.
The game starts with a menu open with a button to play the game. If the player presses "p" the game will be paused. It can be resumed if the button is pressed again.